+++ dartagnan/src/main/java/com/dat3m/dartagnan/expression/processing/ExpressionVisitor.java
@@ -2,10 +2,7 @@ package com.dat3m.dartagnan.expression.processing;
 
 import com.dat3m.dartagnan.expression.*;
 import com.dat3m.dartagnan.program.Register;
-import com.dat3m.dartagnan.program.expression.BinaryBooleanExpression;
-import com.dat3m.dartagnan.program.expression.Comparison;
-import com.dat3m.dartagnan.program.expression.Literal;
-import com.dat3m.dartagnan.program.expression.UnaryBooleanExpression;
+import com.dat3m.dartagnan.program.expression.*;
 import com.dat3m.dartagnan.program.memory.Location;
 import com.dat3m.dartagnan.program.memory.MemoryObject;
 
@@ -16,7 +13,7 @@ public interface ExpressionVisitor<T> {
     default T visit(Literal literal) { return null; }
     default T visit(IExprBin iBin) { return null; }
     default T visit(IExprUn iUn) { return null; }
-    default T visit(IfExpr ifExpr) { return null; }
+    default T visit(ConditionalExpression ifExpr) { return null; }
     default T visit(INonDet iNonDet) { return null; }
     default T visit(Register reg) { return null; }
     default T visit(MemoryObject address) { return null; }
